separo el footer en tres componentes
