+++ src/components/footer/FooterRights.jsx
@@ -1,0 +1,15 @@
+function FooterRights() {
+  return(
+    <div className="container text-light">
+      <div className="row">
+        <div className="col-xs-12 col-sm-6 col-md-8 mx-auto">
+          <p className="copyright-text">
+            Copyright &copy; 2022 All Rights Reserved
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
+
+export default FooterRights;
